feat(triggers): support fractional values for all visible timer fields
Widen every visible trigger timer/duration field from int to float64 so
they accept decimals (e.g. 1.5s), matching the refire-cooldown change and
the CH-chain interval precedent. Fields: timer duration (+ per-pattern
override), visible recast cooldown, display threshold, fading-alert
threshold, and overlay-text display duration.

The spell-timer bridge (StartExternal) and both Sink interfaces now take
float64 durations; the spelltimer internal model already stored floats,
so countdown/expiry already carry sub-second precision. Importers no
longer truncate: GINA (ms + float), EQNag, and EQLogParser durations and
"ending soon" thresholds pass through intact, and ParseDurationText
accepts decimal captures.

Display stays clean per the "integer unless fractional" rule: a new
secsLabel() helper renders whole numbers as plain integers ("16") and
only shows a decimal when one exists ("1.5"); overlay countdowns are
unchanged (still ceil/floor to whole seconds). Frontend inputs switch
parseInt -> parseFloat with step="any".

No DB migration: SQLite stores non-integer reals losslessly in the
INTEGER-affinity columns and collapses whole values back to integers.

+++ backend/internal/api/timers.go
@@ -49,7 +49,7 @@ func (h *timerHandler) clear(w http.ResponseWriter, r *http.Request) {
 func (h *timerHandler) startCustom(w http.ResponseWriter, r *http.Request) {
 	var req struct {
 		Name         string          `json:"name"`
-		DurationSecs int             `json:"duration_secs"`
+		DurationSecs float64         `json:"duration_secs"`
 		Alerts       json.RawMessage `json:"alerts,omitempty"`
 		Color        string          `json:"color,omitempty"`
 	}
